Add MNIST data.
This is a tar.gz'd text format.

The text format is very simple - each line represents a data point. The
line is the 784 pixel values, each separated by a comma. Then a colon,
then the expected value (0-9) e.g.

0.345,0.356,0.9875,0.838 ....:9

The first 50,000 are the training data, the next 10,000 are the
validation data. The final 10,000 are the test data.

This data was created by loading the data in
https://github.com/mnielsen/neural-networks-and-deep-learning.git and
then outputting it as text (it's a python pickle in that repo).